+++ test/syntax/binding_on_syntax.test.ts
@@ -2,7 +2,7 @@ import { expect } from 'chai';
 
 import { Binding } from '../../src/bindings/binding';
 import { BindingScopeEnum } from '../../src/constants/literal_types';
-import { interfaces } from '../../src/interfaces/interfaces';
+import type { interfaces } from '../../src/interfaces/interfaces';
 import { BindingOnSyntax } from '../../src/syntax/binding_on_syntax';
 
 describe('BindingOnSyntax', () => {