+++ test/syntax/binding_when_syntax.test.ts
@@ -10,7 +10,7 @@ import {
   TargetTypeEnum,
 } from '../../src/constants/literal_types';
 import { Container } from '../../src/container/container';
-import { interfaces } from '../../src/interfaces/interfaces';
+import type { interfaces } from '../../src/interfaces/interfaces';
 import { Context } from '../../src/planning/context';
 import { Request } from '../../src/planning/request';
 import { BindingWhenSyntax } from '../../src/syntax/binding_when_syntax';